Add of room occupency

+++ applications/self-star/temperatureManagement/exerciceThree/roomOccupancy/src/main/java/fr/liglab/icasa/self/star/temperature/management/exercice/three/room/occupancy/RoomOccupancy.java
@@ -19,6 +19,6 @@ public interface RoomOccupancy {
      *         that there the room is always empty and 1 indicates that the room
      *         is always occupied at the given moment of the day.
      */
-    public double getRoomOccupancy(double minuteOfTheDay, String room);
+    public double getRoomOccupancy(int minuteOfTheDay, String room);
 
 }
